Update README.md
From turbulent lands to peaceful settlement, this Brave New World Comes Alive! Expand your Nation’s and Peoples borders and influence across the map, in times of war and peace. Reform you nations laws and Culture, conquer lands with Cannon and Rocket, declare oneself Emperor or President. The choice is yours, as YOU are the ruler of this land, the one holding the Crown and Scepter. It is your choice to decide, and at this moment, yours alone.
This mod expands upon base unciv and your power as a ruler, via adding new civs, social policies, mechanics, and more. From Hordes of Barbs Rising to simple flegdeling nations being born out of the ashes of another or anothers land, you are incharge.